requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# HTTP client
httpx==0.27.0

# AI/ML
groq==0.11.0

# RAG / Vector Store / Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.7.0

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.9.0
pydantic-settings==2.5.0

# Optional: Database support
# sqlmodel==0.0.22
# sqlalchemy==2.0.35

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0

# Code quality
ruff==0.6.0